factor out server tests w/ test fixture
