+++ tests/RangeHTTPServer_test.py
@@ -2,14 +2,9 @@
 import sys
 sys.path = ['.'] + sys.path
 import RangeHTTPServer
-from RangeHTTPServer import RangeRequestHandler
 
 from nose.tools import *
 from StringIO import StringIO
-from BaseHTTPServer import HTTPServer
-import requests
-import threading
-import time
 
 
 def test_parse_byte_range():
@@ -42,22 +37,3 @@ def test_copy_byte_range():
     outbuffer = StringIO()
     RangeHTTPServer.copy_byte_range(inbuffer, outbuffer)
     eq_('0123456789abcdefghijklmnopqrstuvwxyz', outbuffer.getvalue())
-
-
-def test_server():
-    httpd = [None]  # no other way to set this from inside start_server() :(
-    def start_server():
-        RangeRequestHandler.protocol_version = 'HTTP/1.0'
-        httpd[0] = HTTPServer(('', 8712), RangeRequestHandler)
-        httpd[0].serve_forever()
-
-    server_thread = threading.Thread(target=start_server)
-    server_thread.setDaemon(True)
-    server_thread.start()
-    time.sleep(1.0)
-
-    r = requests.get('http://localhost:8712/tests/data.txt')
-    eq_('0123456789abcdef\n', r.text)
-
-    httpd[0].shutdown()
-    server_thread.join()
